resuelve conflictos

+++ src/main/java/com/crm/siigroup/crmApp/dto/in/CustomerIn.java
@@ -73,7 +73,7 @@ public class CustomerIn {
     private String addressAditional;
 
     @JsonProperty
-    private int income;
+    private Long income;
 
     @JsonProperty
     private Long currency;
