+++ src/main/java/com/crm/siigroup/crmApp/models/entity/Customers.java
@@ -83,7 +83,7 @@ public class Customers implements Serializable {
     private String addressAditional;
 
     @Column
-    private int income;
+    private Long income;
 
     @ManyToOne
     @JoinColumn(name = "currency_id")
